Restructure project
This commit divides the code into an executable script in bin/ and functionality
in a package dir.

Since relative imports are seriously broken in python, this commit continues to
use absolut imports (as recommended by [PEP
328](http://www.python.org/dev/peps/pep-0328/#rationale-for-absolute-imports)).
This means, however, that you will have to add the `linguini` package dir to
your `PYTHONPATH` when executing `bin/linguini`, for the module needs to be on
python's search path in order to be found when imported absolutely.