Introduce setup.py
This commit introduces a setup script that installs the package.

It moves the executable into the `PATH`, so it may be run without specification
of its location, and moves the module into the python path, so it may be
imported absolutely.

Since the setup script creates a build directory, this commit updates the
.gitignore file.