Generate valid LaTeX documents
So far `linguini` only generated `\include`able recipe snippets.

This commit adds a document template that surrounds the generated snippet with a
preamble and a document environment in order to make it a valid LaTeX document
all by itself.

The ability to generate includable snippets will be reintroduced at a later
point.