Capture STDIN in dedicated process
This commit introduces a dedicated `Process` that captures `STDIN` and a
`Thread` that asynchronously handles the results.

The new process my by alarmed if input was caught from the clipboard, which
leads to the call to `raw_input()` to be interrupted so that the history may be
updated with the external input. In order to be able to capture `STDIN` outside
the main process, the `GTK` event loop that is automatically started upon import
of the respective module needs to be deactivated (`set_interactive(False)`)
until the capturing `Process` has setup and opened `STDIN`.

The new Thread handles input captured by the new process. It blocks until it
reads new input from a `SimpleQueue`. This way the main thread is prevented from
blocking and may receive further updates from the clipboard. In order to
introduce a new thread, the usage of threads must be made known to `GTK` by
calling `threads_init()`.