Eliminate blocking `StdinHandler` thread
So far a `Thread` permanently blocked while reading from the `Queue` that is
filled by the `StdinCapture` process if user input is captured. This thread then
called `process_input()`.

This commit removes said thread and makes the `StdinCapture` process trigger
processing of the input it `put()` into the queue by sending a `SIGALRM` to its
parent. The corresponding handler in the handler then `get()`s the captured
input and causes it being processed.